+++ MSAL/src/instance/MSALAADAuthority.m
@@ -133,7 +133,9 @@
             return @"login.microsoftonline.de";
         case MSALAzureUsGovernmentCloudInstance:
             return @"login.microsoftonline.us";
-            
+        case MSALAzureFranceCloudInstance:
+            return @"login.sovcloud-identity.fr";
+
         default:
             return nil;
     }