+++ Package.swift
@@ -1,11 +1,11 @@
-// swift-tools-version:5.7
+// swift-tools-version:5.9
 
 import PackageDescription
 
 let package = Package(
   name: "MSAL",
   platforms: [
-        .macOS(.v11),.iOS(.v16)
+        .macOS(.v11),.iOS(.v16),.visionOS(.v1)
   ],
   products: [
       .library(
@@ -13,6 +13,6 @@ let package = Package(
           targets: ["MSAL"]),
   ],
   targets: [
-      .binaryTarget(name: "MSAL", url: "https://github.com/AzureAD/microsoft-authentication-library-for-objc/releases/download/2.8.2/MSAL.zip", checksum: "525c9a6a7c4ed04ff647455835eec5e576136107b83f9168fedc6a1362659c35")
+      .binaryTarget(name: "MSAL", url: "https://github.com/AzureAD/microsoft-authentication-library-for-objc/releases/download/2.11.0/MSAL.zip", checksum: "3631e345dfb7b2b923d8b9b4fc172d9bfe15bf1e50b716cdf875b7ea46179a5c")
   ]
 )